+++ project/printf/printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: shachowd <shachowd@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 09:38:21 by shachowd          #+#    #+#             */
-/*   Updated: 2024/07/07 16:25:35 by shachowd         ###   ########.fr       */
+/*   Updated: 2024/07/08 16:34:12 by shachowd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,7 @@
 
 # include <stdarg.h>
 # include <unistd.h>
-# include <stdlib.h>
-# include <stddef.h>
-# include <stdint.h>
+
 
 # if defined(__linux__)
 #  define PTR_NULL "(nil)"
@@ -35,7 +33,7 @@ int		ft_print_percent(void);
 int		ft_print_str(char *str);
 int		ft_print_nbr(int n);
 int		ft_print_unsign(unsigned int nbr);
-int		ft_print_ptr(unsigned long long ptr);
+int		ft_print_ptr(unsigned long ptr);
 char	*ft_strchr(const char *s, int c);
 int		ft_print_hex(unsigned long n, char *strhex);
 int		ft_printf(const char *str, ...);
